created html(s) for doctors/views.py; adjusted placement of all the models
