+++ belasheshe/doctors/templates/doctors/createAppointment.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Create Appointment</title>
+    <!-- Add Bootstrap CSS -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+</head>
+<body>
+    <div class="container mt-4">
+        <h2>Create Appointment</h2>
+
+        <form id="appointmentForm" method="post">
+            {% csrf_token %}
+            <div class="form-group">
+                <label for="member_id">Member ID:</label>
+                <input type="text" class="form-control" id="member_id" name="member_id" required>
+            </div>
+            <div class="form-group">
+                <label for="time">Time:</label>
+                <input type="text" class="form-control" id="time" name="time" required>
+            </div>
+            <div class="form-group">
+                <label for="date">Date:</label>
+                <input type="text" class="form-control" id="date" name="date" required>
+            </div>
+            <button type="submit" class="btn btn-primary">Create Appointment</button>
+        </form>
+    </div>
+
+    <!-- Include Bootstrap and jQuery scripts -->
+    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+
+    <script>
+        $(document).ready(function() {
+            // Submit the form via AJAX
+            $('#appointmentForm').on('submit', function(event) {
+                event.preventDefault();
+                $.ajax({
+                    type: 'POST',
+                    url: '{% url "create_appointment" %}',  // Update with your URL pattern name
+                    data: $(this).serialize(),
+                    success: function(response) {
+                        // Handle success, e.g., show a success message or redirect to another page
+                        alert('Appointment created successfully!');
+                    },
+                    error: function(error) {
+                        // Handle errors, e.g., display error messages
+                        alert('An error occurred. Please try again.');
+                    }
+                });
+            });
+        });
+    </script>
+</body>
+</html>
