+++ belasheshe/doctors/templates/doctors/dashboard.html
@@ -1,45 +1,106 @@
-{% load static %}
 <!DOCTYPE html>
-<html>
+<html lang="en">
 <head>
-    <title>Doctor Dashboard</title>
-    <link rel="stylesheet" type="text/css" href="{% static 'doctors/dashboard.css' %}">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Doctor's Dashboard</title>
+    <!-- Add Bootstrap CSS -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
 </head>
 <body>
-    <div class="dashboard-container">
-        <!-- Doctor dashboard content here -->
-        <header>
-            <h1>Doctor Dashboard</h1>
-        </header>
-
-        <div class="content">
-            <div class="card">
-                <h2>Upcoming Appointments</h2>
-                <ul class="appointment-list">
-                    <li>
-                        <div class="patient-name">John Doe</div>
-                        <div class="appointment-time">August 25, 2023 - 10:00 AM</div>
-                    </li>
-                    <li>
-                        <div class="patient-name">Jane Smith</div>
-                        <div class="appointment-time">August 26, 2023 - 02:30 PM</div>
-                    </li>
-                </ul>
-            </div>
-            <div class="card">
-                <h2>Medical Reports</h2>
-                <ul class="report-list">
-                    <li>
-                        <div class="report-name">Patient: John Doe</div>
-                        <div class="report-date">Reported On: August 20, 2023</div>
-                    </li>
-                    <li>
-                        <div class="report-name">Patient: Jane Smith</div>
-                        <div class="report-date">Reported On: August 22, 2023</div>
-                    </li>
-                </ul>
-            </div>
+    <!-- Navbar -->
+    <nav class="navbar navbar-expand-lg navbar-light bg-light">
+        <a class="navbar-brand" href="{% url 'doctor_dashboard' %}">Doctor's Dashboard</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNav">
+            <ul class="navbar-nav">
+                <li class="nav-item">
+                    <a class="nav-link" href="{% url 'add_prescription' %}">Add Prescription</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="{% url 'create_appointment' %}">Create Appointment</a>
+                </li>
+            </ul>
         </div>
+    </nav>
+
+    <div class="container mt-4">
+        <h2>Welcome, Doctor!</h2>
+
+        <!-- Appointments Section -->
+        <section class="mb-4">
+            <h3>Appointments</h3>
+            <table class="table table-bordered">
+                <thead>
+                    <tr>
+                        <th>Date</th>
+                        <th>Time</th>
+                        <th>Patient Name</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {% for appointment in appointments %}
+                        <tr>
+                            <td>{{ appointment.Date }}</td>
+                            <td>{{ appointment.Time }}</td>
+                            <td>{{ appointment.Member.Name }}</td>
+                        </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+        </section>
+
+        <!-- Checkups Section -->
+        <section>
+            <h3>Checkups</h3>
+            <table class="table table-bordered">
+                <thead>
+                    <tr>
+                        <th>Type</th>
+                        <th>Risk Rate</th>
+                        <th>Details</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {% for checkup in checkups %}
+                        <tr>
+                            <td>{{ checkup.type }}</td>
+                            <td>{{ checkup.riskrate }}</td>
+                            <td>
+                                <button class="btn btn-primary view-details" data-toggle="modal" data-target="#detailsModal{{ forloop.counter }}">View Details</button>
+                                <!-- Modal for checkup details -->
+                                <div class="modal fade" id="detailsModal{{ forloop.counter }}" tabindex="-1" role="dialog" aria-labelledby="detailsModalLabel" aria-hidden="true">
+                                    <div class="modal-dialog" role="document">
+                                        <div class="modal-content">
+                                            <div class="modal-header">
+                                                <h5 class="modal-title" id="detailsModalLabel">Checkup Details</h5>
+                                                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                                    <span aria-hidden="true">&times;</span>
+                                                </button>
+                                            </div>
+                                            <div class="modal-body">
+                                                <p>Type: {{ checkup.type }}</p>
+                                                <p>Risk Rate: {{ checkup.riskrate }}</p>
+                                            </div>
+                                            <div class="modal-footer">
+                                                <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </td>
+                        </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+        </section>
     </div>
+
+    <!-- Include Bootstrap and jQuery scripts -->
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 </body>
 </html>
