+++ belasheshe/doctors/views.py
@@ -1,67 +1,66 @@
 from django.shortcuts import render, redirect
 from django.views import View
-from residents.models import Member, MemberAppoint, MedicineChart,Medicine,Dosage
-from nurse.models import SpecialCheckupSchedule
-from .models import Doctor
+from nurse.models import MemberAppoint, MedicineChart,Medicine,Dosage, SpecialCheckupSchedule, CheckupSchedule
+from residents.models import CheckupItem
+# from .models import Doctor
 
 # Create your views here.
 # def doctor_dashboard(request):
 #     return render(request, 'doctors/dashboard.html')
 
-def member_medical_conditions(request):
-    # Retrieve all members with associated medical conditions and current conditions
-    members = Member.objects.all()
+# def member_medical_conditions(request):
+#     # Retrieve all members with associated medical conditions and current conditions
+#     members = Member.objects.all()
     
-    # Sort members based on Risk_rate (ascending order)
-    members = members.order_by('currentcond__Risk_rate')
+#     # Sort members based on Risk_rate (ascending order)
+#     members = members.order_by('currentcond__Risk_rate')
 
-    # Paginate the members list (adjust the items per page as needed)
-    paginator = Paginator(members, 10)  # Show 10 members per page
-    page_number = request.GET.get('page')
-    page_members = paginator.get_page(page_number)
+#     # Paginate the members list (adjust the items per page as needed)
+#     paginator = Paginator(members, 10)  # Show 10 members per page
+#     page_number = request.GET.get('page')
+#     page_members = paginator.get_page(page_number)
 
-    medical_conditions = []
+#     medical_conditions = []
 
-    for member in page_members:
-        # Retrieve the member's general information
-        member_info = {
-            'name': member.Name,
-            'address': member.Address,
-            'room_no': member.Room_no,
-            'email': member.Email,
-            'phone': member.Phone,
-            'dob': member.DOB,
-            'religion': member.Religion,
-        }
+#     for member in page_members:
+#         # Retrieve the member's general information
+#         member_info = {
+#             'name': member.Name,
+#             'address': member.Address,
+#             'room_no': member.Room_no,
+#             'email': member.Email,
+#             'phone': member.Phone,
+#             'dob': member.DOB,
+#             'religion': member.Religion,
+#         }
 
-        # Retrieve the medical conditions for each member
-        med_conditions = ResidentMedCond.objects.filter(Member_ID=member.Member_ID)
-        med_condition_info = []
+#         # Retrieve the medical conditions for each member
+#         med_conditions = ResidentMedCond.objects.filter(Member_ID=member.Member_ID)
+#         med_condition_info = []
 
-        for med_cond in med_conditions:
-            # Retrieve the details of each medical condition
-            condition = MedCond.objects.get(Cond_ID=med_cond.Cond_ID)
-            med_condition_info.append({
-                'condition_name': condition.Name,
-                'advice': condition.Advice,
-                'guideline': condition.Guideline,
-            })
+#         for med_cond in med_conditions:
+#             # Retrieve the details of each medical condition
+#             condition = MedCond.objects.get(Cond_ID=med_cond.Cond_ID)
+#             med_condition_info.append({
+#                 'condition_name': condition.Name,
+#                 'advice': condition.Advice,
+#                 'guideline': condition.Guideline,
+#             })
 
-        # Retrieve the current condition for each member
-        try:
-            current_condition = CurrentCond.objects.get(Member_ID=member.Member_ID)
-            risk_rate = current_condition.Risk_rate
-        except CurrentCond.DoesNotExist:
-            risk_rate = None
+#         # Retrieve the current condition for each member
+#         try:
+#             current_condition = CurrentCond.objects.get(Member_ID=member.Member_ID)
+#             risk_rate = current_condition.Risk_rate
+#         except CurrentCond.DoesNotExist:
+#             risk_rate = None
 
-        medical_conditions.append({
-            'member_info': member_info,
-            'conditions': med_condition_info,
-            'risk_rate': risk_rate,
-        })
+#         medical_conditions.append({
+#             'member_info': member_info,
+#             'conditions': med_condition_info,
+#             'risk_rate': risk_rate,
+#         })
 
-    return render(request, 'doctors/medical_conditions.html', {'medical_conditions': medical_conditions, 'page_members': page_members})
-
+#     return render(request, 'doctors/medical_conditions.html', {'medical_conditions': medical_conditions, 'page_members': page_members})
 
 
 class DashboardView(View):
@@ -100,11 +99,14 @@ class DashboardView(View):
     def get(self, request, doctor_id, date, *args, **kwargs):
         appointments = self.getAppointments(doctor_id)
         checkups=self.getCheckups(doctor_id)
+    # def get(self, request, *args, **kwargs):  # testing
+    #     appointments =[]
+    #     checkups=[]
         return render(request, self.template_name, {'appointments': appointments, "checkups": checkups})
 
 class addSuggesions(View):
     template_name = 'doctors/addPrescription.html'  # Update with your actual template name
-    def get(self, request, doctor_id, date, *args, **kwargs):
+    def get(self, request, *args, **kwargs):
         return render(request, self.template_name)
 
     def addPrescriptopn(self, request, member_id):
@@ -121,7 +123,7 @@ class addSuggesions(View):
             new_instance = MedicineChart(Member_id=member_id, Date=date)
             new_instance.save()
 
-            chart_id=new_instance.Chart_id;
+            chart_id=new_instance.Chart_id
 
             instance = Dosage(Chart_id=chart_id, Medicine_id=medicine_id, Time=time, Quantity=quantity)
             instance.save()
@@ -129,7 +131,7 @@ class addSuggesions(View):
 
 class createAppointment(View):
     template_name = 'doctors/createAppointment.html'  # Update with your actual template name
-    def get(self, request, doctor_id, date, *args, **kwargs):
+    def get(self, request, *args, **kwargs):
         return render(request, self.template_name)
 
     def addAppointment(self, request, doctor_id,member_id):
