+++ belasheshe/nurse/templates/nurse/checkup_data_list.html
@@ -4,26 +4,27 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Checkup Data List</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous"></head>
+    <!-- Add Bootstrap CSS -->
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+</head>
 <body>
     <div class="container mt-4">
         <h2>Checkup Data List</h2>
 
         <!-- Sorting options -->
         <div class="mb-3">
-            <label for="sortSelect">Sort by:</label>
-            <select id="sortSelect" class="form-control">
+            <label for="sortSelect" class="form-label">Sort by:</label>
+            <select class="form-select" id="sortSelect">
                 <option value="blood_pressure">Blood Pressure</option>
                 <option value="sugar">Sugar</option>
                 <option value="heartrate">Heart Rate</option>
             </select>
         </div>
-        
-        <!-- Display sorted checkup data -->
+
         <table class="table table-bordered">
             <thead>
                 <tr>
-                    <th>Member ID</th>
+                    <th>Member Name</th>
                     <th>Blood Pressure</th>
                     <th>Sugar</th>
                     <th>Heart Rate</th>
@@ -32,26 +33,45 @@
             <tbody>
                 {% for data in checkup_data %}
                     <tr>
-                        <td>{{ data.member.Member_ID }}</td>
-                        <td>{{ data.item.BloodPressure }}</td>
-                        <td>{{ data.item.Sugar }}</td>
-                        <td>{{ data.item.Heartrate }}</td>
+                        <td>{{ data.Member_Id.Name }}</td>
+                        <td>{{ data.Checkup_id.Blood_Pressure }}</td>
+                        <td>{{ data.Checkup_id.Sugar }}</td>
+                        <td>{{ data.Checkup_id.Heartrate }}</td>
                     </tr>
                 {% endfor %}
             </tbody>
         </table>
+
+        <!-- Pagination links -->
+        <div class="pagination">
+            <span class="step-links">
+                {% if checkup_data.has_previous %}
+                    <a href="?page=1{% if order_by %}&order_by={{ order_by }}{% endif %}">First</a>
+                    <a href="?page={{ checkup_data.previous_page_number }}{% if order_by %}&order_by={{ order_by }}{% endif %}">Previous</a>
+                {% endif %}
+
+                <span class="current-page">{{ checkup_data.number }}</span>
+
+                {% if checkup_data.has_next %}
+                    <a href="?page={{ checkup_data.next_page_number }}{% if order_by %}&order_by={{ order_by }}{% endif %}">Next</a>
+                    <a href="?page={{ checkup_data.paginator.num_pages }}{% if order_by %}&order_by={{ order_by }}{% endif %}">Last</a>
+                {% endif %}
+            </span>
+        </div>
     </div>
 
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-    
+    <!-- Include Bootstrap and jQuery scripts -->
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.3/dist/umd/popper.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+
     <script>
         $(document).ready(function() {
-            // Handle user selection of sorting field
-            $('#sortSelect').on('change', function() {
-                const selectedValue = $(this).val();
-                window.location.href = `/checkup_data/?order_by=${selectedValue}`;
+            // Handle sorting option change
+            $('#sortSelect').change(function() {
+                var selectedSort = $(this).val();
+                var currentUrl = window.location.href.split('?')[0];  // Get the current URL without query parameters
+                window.location.href = currentUrl + '?order_by=' + selectedSort;  // Redirect with the selected sorting option
             });
         });
     </script>
